extracted CombineLatestOperators class. Put Javadocs.
