+++ src/main/java/org/ossgang/commons/observables/operators/OperatorUtils.java
@@ -1,0 +1,28 @@
+package org.ossgang.commons.observables.operators;
+
+import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.IntStream;
+
+import static java.util.stream.Collectors.toList;
+import static java.util.stream.Collectors.toMap;
+
+public final class OperatorUtils {
+
+    static <V> Map<Integer, V> toIndexMap(List<V> list) {
+        return IntStream.range(0, list.size()).boxed().collect(toMap(Function.identity(), list::get));
+    }
+
+    static <V> List<V> fromIndexMap(Map<Integer, V> map) {
+        return IntStream.range(0, map.size()).boxed().map(map::get).collect(toList());
+    }
+
+    static <K, I, O> Map<K, O> applyToMapValues(Map<K, I> inputMap, Function<I, O> mapper) {
+        return inputMap.entrySet().stream().collect(toMap(Map.Entry::getKey, e -> mapper.apply(e.getValue())));
+    }
+
+    private OperatorUtils() {
+        throw new UnsupportedOperationException("static only");
+    }
+}
